% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripp.R
\name{ripp.p.outcome}
\alias{ripp.p.outcome}
\title{Partial identification with imperfect placebo outcome}
\usage{
ripp.p.outcome(lm.y.dpx, lm.n.dpx, treatment)
}
\arguments{
\item{lm.y.dpx}{\code{lm(Y~D+P+X)} object.}

\item{lm.n.dpx}{\code{lm(N~D+P+X)} object.}

\item{treatment}{The name of the actual treatment.}
}
\description{
\code{Y} is the actual outcome. \code{N} is the placebo outcome. \code{D} is the treatment.
}
