% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripp.R
\name{ripp.p.treatment}
\alias{ripp.p.treatment}
\title{Partial identification with imperfect placebo treatment}
\usage{
ripp.p.treatment(lm.y.dpx, treatment, placebo_treatment)
}
\arguments{
\item{lm.y.dpx}{\code{lm(Y~D+P+X)} object. \code{Y} is actual outcome.}

\item{treatment}{The name of the actual treatment.}

\item{placebo_treatment}{\code{P} is placebo treatment.}
}
\description{
Partial identification with imperfect placebo treatment
}
