% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripp.R
\name{ripp}
\alias{ripp}
\title{Partial identification with imperfect placebos}
\usage{
ripp(type = c("placebo outcome", "placebo treatment", "double placebo"), ...)
}
\arguments{
\item{type}{"outcome", "treatment", or "double"}

\item{...}{arguments passed to other methods. First argument should be a \code{\link{lm}} with the
outcome regression (argument \code{lm.y.dpx} of the form \code{lm(Y~D+P+X)}). Second argument
should be a \code{\link{lm}} with the placebo outcome regression (argument \code{lm.n.dpx} of the form \code{lm(N~D+P+X)}),
when a placebo outcome, \code{N}, is available. \code{P} is a placebo treatment that is included in the regression for
\code{Y} when a placebo treatment is available. When both a placebo outcome and a placebo treatment are available,
\code{P} is also included in the regression for \code{N}.}
}
\value{
\code{ripp} object that includes:
}
\description{
This function performs partial identification (or sensitivity analysis) of causal
effects leveraging information about imperfect placebo outcomes and/or treatments.
It returns an object of class \code{ripp} that contains key metrics for partial identification.
}
