% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripp.R
\name{ripp.double.p}
\alias{ripp.double.p}
\title{Partial identification with imperfect placebo outcome and placebo treatment}
\usage{
ripp.double.p(lm.y.dpx, lm.n.dpx, treatment, placebo_treatment)
}
\arguments{
\item{lm.y.dpx}{\code{lm(Y~D+P+X)} object. \code{Y} is actual outcome.}

\item{lm.n.dpx}{\code{lm(N~D+P+X)} object. \code{N} is placebo outcome.}

\item{treatment}{The name of the actual treatment.}

\item{placebo_treatment}{\code{P} is placebo treatment.}
}
\description{
Partial identification with imperfect placebo outcome and placebo treatment
}
