% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripp.R
\name{ripp.p.outcome}
\alias{ripp.p.outcome}
\title{Partial identification with imperfect placebo outcome}
\usage{
ripp.p.outcome(lm.y.dpx, lm.n.dpx, treatment)
}
\arguments{
\item{lm.y.dpx}{\code{lm(Y~D+P+X)} object. \code{Y} is actual outcome.}

\item{lm.n.dpx}{\code{lm(N~D+P+X)} object. \code{N} is placebo outcome.}

\item{treatment}{The name of the actual treatment.}
}
\description{
Partial identification with imperfect placebo outcome
}
